Disable metric pausing during warmup
Fixes: 152387965
Test: ./gradlew ui:integration-tests:benchmark:cC -Pandroid.testInstrumentationRunnerArguments.class=androidx.ui.core.OnPositionedBenchmark#deepHierarchyOnPositioned_layout # now runs in ~17 seconds instead of many minutes

Timing for warmup happens for two reasons:

1) looking for stabilization of the workload
2) estimation of how long the workload will take.

Prior to the warmup rearchitecture, warmup timings (somewhat
coincidentally) didn't respect pause/resume, since WarmupManager never
observed those times.

This does mean that we're looking at stabilization of the entire loop,
instead of skipping the runWithTimingDisabled portion. This is
somewhat unfortunate, but it has two bigger implications:

First, our warmup max budget of 8 seconds is always respected. If a
benchmark loop is 1% measured, we don't want to have to run 800
seconds to fulfill 8 seconds of true warmup.

Second, our estimation of number of times we should run the loop. For
similar reasons, if we expect to run a benchmark for 0.5ms per repeat,
we don't want to run 500ms to get that 0.5ms of execution of the
measured section.

Since we saw exactly these problems in CI (tests timing out due to
massively increased runtimes), this change reverts us back to the
previous behavior of skipping pause/resume during warmup.

Additionally, we add tracing to runWithTimingDisabled, so it's easier
to see when a test is 1% measured (since it happens in practice)

Change-Id: I04eda610a7c896e089a29e9835f03e87a5bc5be9