Unwrap + safe cast the app context to Application
Since getApplicationContext() returns a Context and
not an Application, developers can use one or more
ContextWrappers to provide any number of layers of
interdirection between the Context returned and the
actual Application instance.

By calling getBaseContext() until we find an
Application instance and providing a fallback if
we never find one, we can ensure that we always
get a valid default factory.

Test: existing ViewModel tests still pass
BUG: 153762914
Change-Id: If9b6ecd25db2d8acebb4d1a72baae4b3510f5a7e