Measure with layout direction in Layout DSL
The CL updates the DSL used for Layout and LayoutModifier(2) to be able
to measure children with a different layout direction. The CL also fixes
automirroring of positions for the wrapped layout of LayoutModifier2,
and updates the ltr and rtl modifiers to use LayoutModifier2.
API for changing layout direction in intrinsic measurements will come in
a following CL.

Relnote: Layout and LayoutModifier children can be measured with a different layout direction.
Test: ran ui-layout, ui-framework tests and added new test in RtlLayoutTest
Bug: 153971756
Bug: 153317665
Bug: 153310100
Change-Id: Ibd7319b7caa93b2bc7fb38df3678e2bb8298652e