Fix LoadState updates from RemoteMediator
doBoundaryLoad now correctly acquires the stateLock before updating
LoadState from RemoteMediator result.

RemoteMediator.loads of type REFRESH now correctly updates LoadState to
Loading. It's a little bit unfortunate, but canRequestMoreData is
ignored for REFRESH currently as it's not valid for REFRESH to be set to
LoadState.Done, which would prevent loading of any data.

Note that the RemoteMediator LoadState behavior for REFRESH diverges
slightly from START / END as it does not flip between Idle and Loading
when finishing the Remote call and triggering PagingSource.load. This
helps to avoid flickering of loading spinner for REFRESH as we are
always guaranteed PagingSource.Load(REFRESH) to be called after a
successful RemoteMediator.load(REFRESH).

Test: ./gradlew paging:paging-common:test
Change-Id: I550ca035842e0aeed49f9fb2a58a25fad316a055