Add "clear" methods to FragmentResult APIs
Currently the FragmentResultOwner class only has the setFragmentResult()
and setFragmentResultListener methods. Both of these methods take a
Nullable parameter and in order to clear the result/listener developers
need to pass in null for that parameter. While this does work, it would
be more explicit if there were a separate API.

This change adds clearFragmentResult() and clearFragmentResultListener()
as the method of removing previously set results/listeners. The set
methods above no longer take Nullable parameters.

This change also adds the clear method to the Fragment -ktx.

Test: FragmentResultTest
Bug: 155416778
Change-Id: I3dfeb38a319829c6305e683b0bf335b92daaaa46