Rewrite old ActivityResult methods in Fragment
With the release of the new ActivityResult APIs, Fragment can now take
advantage of those instead of using the framework method of handling
starting activities for results, or requesting permissions.

Using the ActivityResult APIs means there is no longer a requirement for
requestCodes to only use the lower 16 bits, and NestedFragmentTest have
been adjusted accordingly.

Test: existing test pass
Bug: 155518741
Change-Id: Ifddbbd362e2e57cb9cdd30f9e19626c0c9661384