Call noteStateNotSaved in FragmentActivity
We used to call noteStateNotSaved in FragmentActivity, which allowed any
subclasses of FragmentActivity to assume that the savedState was cleared
in the onActivityResult callback, even if onActivityResult was never
actually dispatched to any Fragment. This means the state was always
cleared when you got on the onActivityResult() of the subclass, and
therefore it was possible to do fragmentTransactions.

We should override the onActivityResult() and
onRequestPermissionsResult() methods and call noteStateNotSaved to
remain consistent with previous behavior.

Test: FragmetnReceiveResultTest
Bug: 156271436
Change-Id: Ie3b9156b861d042493ab76e97e6a9af1a14a1e8d