AndroidXHiltProcessor no longer extends BasicAnnotationProcessor.
BasicAnnotationProcessor forces annotations in runtime classpath be
mixed with compile classpath and this is an issue if the annotations
are in an Android packaged library (aar), see b/153170868.

Since Hilt is now being correctly built in an AAR we can no longer
depend on it on our java-only common artifact, therefore make that
library an android library module and refactor the compiler to not
depend on it.

Test: ./gradlew hilt:hilt-compiler:test
Change-Id: I66aa2c70deda3297b43807d60d713436dd3f57d9