Ensure the layout tree is measured before handling the touch events
It is possible that dispatchTouchEvent could be called after the composition but before we remeasure the LayoutNodes to apply the final changes. It is essential that touch processing is operating over the measured nodes so we should manually call measureAndLayout() before processing similarly how we do so in dispatchDraw(). This method does nothing when there is nothing scheduled to remeasure so it will not make it slower

Test: covered with a new test
Fixes: 156370792
Change-Id: I8067dab7e0d1838dc86dcd0a7e9043cfe06d351c