Poke Android when a layer moves.
When a layer moves, Android requires a nudge to indicate
that something needs to be redrawn. This CL gives that nudge
with a View invalidate(), which should do minimal drawing
effort since the root is in its own layer and doesn't need
to be re-recorded.

The same is true for all other types of simple RenderNode
property changes -- they need to trigger a redraw.

Fixes: 156950983
Test: demo app, new test
Change-Id: I7b6a389be8d9d0bd04d4286e5f027924a0e02ce8
Merged-In: I7b6a389be8d9d0bd04d4286e5f027924a0e02ce8