Updated pointer input bounds behavior.
Pointer input used to be clipped by all layout bounds
such that pointer input fitlers could only be interacted
with if they were in a region that intersected all of the
regions of layout bounds above it.

This behavior has been removed and replaced with
automatic pointer input clipping when draw clipping is used.

Fixes: 155868092
Test: ./gradlew ui:ui-core:test
Test: ./gradlew ui:ui-core:connectedCheck
Test: ./gradlew ui:ui-core:testDebugUnitTest --tests=androidx.ui.core.pointerinput.PointerInputEventProcessorTest
Test: ./gradlew ui:ui-core:testDebugUnitTest --tests=androidx.ui.core.LayoutNodeTest
Test: ./gradlew ui:ui-core:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.ui.core.test.ClipDrawTest
Test: ./gradlew ui:ui-core:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.ui.core.test.ClipPointerInputTest

Change-Id: I356255464eb90d49f82fe5c41c6a39aed14baa0a