Add FragmentHostCallback optional interface information
Add information on FragmentHostCallback on
optional interfaces that a FragmentHostCallback
can implement to change the behavior of
FragmentManager and remove the need to use
deprecated APIs on FragmentHostCallback and
FragmentController.

Also adds the missing Javadoc on the <E> template
class to clarify that it is the type of object
that is hosting the fragments.

Test: ./gradlew bOS
Change-Id: Ib1ccdbbcd28acc55ec3dd5a469d7bd328499a503