Add multi-touch gesture injection APIs for tests
This adds the ability to specify pointer ids to the partial gesture
methods in PartialGestureScope, and reduces the API in InputDispatcher
to the bare minimum needed to serve all gestures. All prefab gestures in
InputDispatcher now use the methods to dispatch individual motion events
as well.

The test suite has gotten a major overhaul, trying to make them simpler
while still testing the right thing. This means less parameterized
tests, and more test methods.

Bug: 145593752
Bug: 157142514
Test: ./gradlew ui:ui-test:cC
Relnote: "
Added multi-touch gesture injection APIs to be able to test
components when performing multi-touch gestures.

When injecting a gesture, one now has the ability to specify multiple
pointers by supplying a pointer id to a partial gesture method. If no
pointer id is specified, the pointer 0 is assumed, called the default
pointer. Note that all prefab gestures (`sendClick`, `sendSwipe` and
all methods built on these) use the default pointer. The following
methods in PartialGestureScope now have an overload that accepts a
pointer id:
- `sendDown`
- `sendMoveTo`
- `sendMoveBy`
- `sendUp`

Each event now contains a complete list of the current location of each
active pointer (an active pointer being one that is down), and the
methods listed above send an event immediately. In addition, three new
methods have been introduced:
- `movePointerTo`
- `movePointerBy`
- `sendMove`

The methods `movePointerTo` and `movePointerBy` change the current
location of a pointer, but do not send an event to reflect that.
Consequently, they also don't advance the current time of the gesture.
This allows one to move several pointers simultaneously. The last
method, `sendMove`, does not accept any parameters and simply sends a
move event, allowing one to \"flush\" all pointer movement. If pointers
have been updated with `movePointerXx`, and a down or up event is sent,
a move event will be sent immediately preceding the down or up event.

The timing of all methods has slightly changed as well. As of now,
`sendDown` and `sendUp` no longer advance the time of the gesture,
so one can send multiple pointers down or up simultaneously. All the
other methods do advance the time of the gesture. It is currently not
(yet) possible to change the time delta.
"

Change-Id: I9acaccaaf659868c78de6a1b13f1c4acb8979836