Return an empty list from OpenMultipleDocuments
The OpenMultipleDocuments contract returns a null List<Uri> when the
startActivityResult call fails. We should instead return an empty list.
This allows us to be consistent with the GetMultipleContents contract
and the empty list is nicer to work with.

Test: ./gradlew checkApi
Bug: 157348014
RelNote: "The OpenMultipleDocuments contract now returns an empty list
instead of null to the registered callback if no result is returned"

Change-Id: Ie4c991bfcbddc17811a37b226dec3f765f9813f1