Changes in captureToBitmap.
Fixes the captureToBitmap to work with ActionBar and cut outs. This also
replaces the way we were obtaining the screenshots. Instead of taking a
bitmap from a window we use UiAutomator which actually takes a real
screenshot. This enabled us to capture multiple windows and defend
against cases where popup could be on top of our view. Finally it also
removes the need to juggle with Activities (and casting them) and also
removes the O+ API restriction.

Finally I have also removed the method used to capture full screen. The
reason is that no one used it and it is tricky to define how it should
work. E.g. it does not have much value if we don't crop out the action
bar. With that and the fact that developers can use UiAutomator for
fullscreen screenshots it is not beneficial to keep that API around. We
can reintroduce it in the future once we need to.

Bug: 157657482
Test: Added
Change-Id: I5320d61c7a4ebc03472b0cb79c3da6c3870482f3