Track entering and exiting Views separately in tests
Rather than using a single varargs list for all
Views involved in transitions, track the set of
entering and exiting Views separately to allow for
easier debugging of test failures.

This switches the verifyAndClearTransition() test API
to using a receiver scope lambda to fill in a data
class on the information to pass through to the test.
This helps only pass the exact set of information
that is needed in a readable format rather than rely on
default parameters of a method.

Test: all existing tests still pass
BUG: 147749580
Change-Id: I072882a75a7393ac0dc4ea6797a6b6a91e50b94f