Add vertical scrolling behaviour to the FilledTextField
This CL is to fix current issue of the input text not being clipped and not being layed out correctly when you restrict the height of the FilledTextField. However, with that change we will not respect anymore the spec related to the offset between the label and the input field.

Test: FilledTextFieldTest with new tests + checked demos
Fixes: b/157210750
Change-Id: Ie3bd87e040b5432356fcf244dbabb74d140e4b71