Fix ConstraintLayout crash on specific remeasure
Before this CL, ConstraintLayout was storing a map of Measurables to
Placeables across remeasurements. Also, based on it, ConstraintLayout
was querying for alignment lines of Placeables before measuring again
the layout. In certain layout scenarios, this could lead to a crash
because of asking for alignment lines of unmeasured layouts. This CL
prevents this, by no longer keeping the map but just the values obtained
during the previous measurement.

Bug: 158454730
Test: ConstraintLayoutTest
Change-Id: I366e40c9ae84bce89bfa65572e8c358ee7ab8c82