Provide a view scoped SavedStateRegistryOwner
Rather than returning the Fragment itself as the
ViewTreeSavedStateRegistryOwner for a Fragment,
create a separate SavedStateRegistryOwner that is tied
into the view lifecycle.

This ensures that the SavedStateRegistryOwner provided
there correctly has its state saved alongside the
View itself and restored when the view state is restored.

This particularly fixes cases where the Fragment
is put on the back stack and the Fragment itself
has not gone through onSaveInstanceState().

Test: new FragmentViewLifecycleTest passes
BUG: 158503763
Change-Id: I3741fc1a1b289d6594b22a192c4496625ec9e5e9