Bugfix: in viewport scale type, start/end should change with display rotation.
The definition of start/end changes with rotation and/or layout direction. Assume for example in Fig.1, the rotation degree of the output is 90, which means the user should rotate it 90 degrees clockwise to correct the output (as seen in Fig.2) . In this case, the "start" of the buffer in Fig.1 is no longer at (0, 0) but (0, w).

                (0, 0)
  +--------+      +--------+
  |        |      |    \   |
  |  /\__  |      |    /   |
  |        |      |   |    |
  +--------+      +--------+
(0, w)

   Fig.1           Fig.2

 This CL corrects the crop rect based on rotation and layout direction. Also added extensive tests to cover all possbile scenarios.

Bug: 153096869
Test: manual test and ./gradlew bOS
Change-Id: I2c334bd388cf3c30e9fd23470d531d3b098d22d6