Fix for AdapterList can't be scrolled when item is initially created with 0 size and then expanded
It happens because we create an item to consume the scroll when we are in the bottom of the list. we measure it and it has 0 size. as we can't consume scroll we don't move. and this item is outside of the visibility so we immediately remove it so it never expands.
It is in general incorrect to use AdapterList like this so we first need to add a comment that even if you don't yet have data to display you should reserve space with Spacer or similar if you are not going to expand right after and second we can handle such case nicer by not removing the items which are right on the edge(their position == list height), which seems reasonable

Fixes: 159032170
Test: manually, plus new test
Change-Id: I7c1a6e10fbe3c9d7a3dc3fc43d5dbf87068c81ad