Tear down MultiInstanceInvalidator syncronously
This CL fixes an issue in MultiInstanceInvalidation where we would post a runnable
to room's executor when database#close is called. I've changed it to be sync since
it is called when close lock is acquired.

I've also changed the test to reserve a database per test name so that tests won't
race eachother when deleting the database file. Normally, this shouldn't be an issue
but we do have some problems where an invalidated live data might be preparing to
read from database when it closes and that messes up things if the test for that
live data already finished. This is a workaround for that where each test has its
own database so that they only try to delete the database from a previous run.

Bug: 152501177
Test: MultiInstanceInvalidationTest (500 times)
Change-Id: I029debdf0962e554db70f9c955fcc74f70644070