MediaRouter: Add onPrepareTransferListener
By adding a listener, developers can defer releasing of the previously
selected route controllers, which can be used to synchronize media
status.
They can also determine when to stop the previous route
to implement "seamless transfer"

It doesn't affect the order of onRouteUnselcted/onRouteSelected
but the timing of RouteController#onReleased.

RelNote: "Add a new listener `MediaRouter#OnPrepareTransferListener`
for receiving events when the selected route is about to be changed"
Bug: 159963791
Test: ./gradlew mediarouter:mediarouter:check &&
    manually using support v7 demos w/ MR2.
    Confirmed we can control the timing of stopping the previous cast.

Change-Id: I6ace162aec4d12c70d8da094041ae13126e6d61f