Merge DeviceState and WindowLayoutInfo
Current implementation is not extensible. If a device with 2 (or more)
fold features is exist, we can not handle two device state.
So merge DeviceState and WindowLayoutInfo callbacks and
introduce a WindowFeature.

Relnote: "Merge DeviceState and WindowLayoutInfo so it is easier to
access data."
Relnote: "Convert DisplayFeature into an interfac to support future
features."

Test: ./gradlew window:window:connectedAndroidTest
Change-Id: Id34f448e9403aa6462a63b574073a6762365ed6f