Replace Types.isAssignable(from, to) with to.isAssignableFrom(from)
Type.isAssinable is clear on which type mirror is assigned to the
other one, making it hard to read the code.

This CL replaces it with an extension function. This is also a
preliminary step for processor abstraction where we wont need to
pass the Types instace, make it read line `to.isAssignableFrom(from)`.

Bug: 160323720
Test: existing test suite
Change-Id: I14dd473b525e979cee93375db925d85a31c7d6f9