Add MetadataTransform API to CameraGraph
MetadataTransform is provided as a way to alter the K/V pairs returned
by camera2 and to make it possible to "shift" keys forward or backwards.

To make this consumable (while still providing fast, direct access to
metadata) this CL changes the request callbacks so that there is a
seperate method to get a corrected version of metadata from the camera.

By default the transform is a NoOp and most consumers should use the
ResultMetadata interface to access information about the frame.

Test: Added unit tests MetadataTransform behavior.
Test: Existing unit tests continue to pass.

Change-Id: Id5ddd69c41c1560591907eb86b00a957422e4ef6