Migrate Field.element's type to VariableElement
It was already always a VariableElement, we just didn't specify it.

This will help us avoid accessing enclosingElement on unknown
types (except for javac specific parts like JVMDescriptorUtils or
KotlinMetadata extraction).

In XProcessor, I'm trying to control the access schema to be one way.
For instance, the only way to get a field is via its owning TypeElement
(or one of its subtypes).
This means, we can pass down kotlin metadata to it (similar for methods).

The only access pattern that goes against the 1 way flow is the
enclosing element checks, which we use to check things like where
an element was declared (important for inherited fields).
Keeping 1 way also helps with caching where we can cache XTypeElement
and then each will cache its methods etc (and also always return them
as member of the element).

I'm not sure how well that will scale as it is to be seen when I try
the migration but I thought this change by itself is good anyways.

Bug: 160323720
Test: existing test suite
Change-Id: Id75518a28bdb05252ed1e504d9627325fc673c31