Don't throw exception if the view isn't attached
The recent changes to WindowInsetsCompat caused a regression
for testing insets on un-attached views (usually done with
obolectric).

This CL makes the code more permissive, and just ignores the
view if it is not attached.

Also marked the IME tests as flaky. Testing the IME is inherently
flaky so there's not a lot we can do.

Test: manual
BUG: 160794833
Change-Id: I7218ed8da8c46d3fd908477b8db20e544c9ed810