Add support for setting the navbar divider color.
Users required the ability to change the color of the navigation bar
divider which looked inconsistent in custom tabs when both the
contents of a web app and nav bar were the same color. This commit
extends the CutomTabColorSchemeParams and
TrustedWebActivityIntentBuilder with a color field for setting the
navbar divider color.

Relnote: "Added the setNavigationBarDividerColor method to
TrustedWebActivityIntentBuilder and CustomTabColorSchemeParams."

Test: Added test that ensure the  navbar divider color can be set and transmitted through an intent bundle. Ran the tests with ./gradlew browser:browser:test

Change-Id: Ia04ddf4f141e6e48ef2e25c88f9f2f101c5b3599