Fix ConcurrentModificationException in tryDetachUseCases
When unbindAll, mCameraUseCaseAdapter.getUseCases() is passed
to CameraUseCaseAdapter#removeUseCases directly and it will
call mCameraInteral.detachUseCases() and then remove the use
cases from CameraUseCaseAdapter.mUseCases.

But the problem is mCameraInteral.detachUseCases will run
tryDetachUseCases in camera executor,  thus it is possible to
cause a race condition that CameraUseCaseAdapter.mUseCases
are used in a for-loop in tryDetachUseCases and are being
modified in adapter at the same time.

Also mCameraUseCaseAdapter.getUseCases() returns the mUseCass
to the other parties which we have no way to protect it using
internal lock,  Thus we shall copy it to a new instance to
avoid possible crash.

Bug: NA
Relnote: "Fixed the ConcurrentModificationException exception
issue when unbinding use cases."
Test: Runs all camera core tests , camera2 tests
Change-Id: I371c052c5edd6aeed2a127a44cc422758417ea57