Return Button modifier argument to outermost position
In
https://android-review.googlesource.com/c/platform/frameworks/support/+/1354246/8/ui/ui-material/src/main/java/androidx/ui/material/Button.kt
, I introduced a bug: I moved the passed-in "modifier" argument
to an inner composable of the Button instead of leaving it on the
surface.  The reason I did that is because I needed to have
modifiers on the same node or below as clickable() so testTags would get
merged into it, and moving clickable() to the top didn't work because
it broke ripple containment.  In this patch I finessed the problem
better by managing the ripple explicitly in the Button instead of
relying on clickable()'s default arguments.

Relatedly, I also broke the ripple in FloatingActionButton in the same
patch, so fix that in the same way.

Test: weightModifierOnButton, fab_weightModifier
Fixes: 161007448
Change-Id: I9a126ae2d3107bfb87e2d2d65f1a9b1e89c5fd61