Add an implementation of RequestProcessor
This class adds an implementation of the RequestProcessor instance that
will create and submit CaptureRequest(s) to a Camera2 CaptureSession.
It also adds an internal CaptureSequence class that forwards capture
events the event listeners that are listed on the Request object.

This implementation does not add tests (yet). It's not possible to
create CaptureRequest, CaptureRequest.Builder or CaptureResult objects
directly, and shadow implementations of these objects are not available
until Robolectric 4.4 (Current version is 4.3.1)

Test: Existing tests continue to pass
Change-Id: I509f0ac4247141850949be4ea25137e1fd0f8e6f