Add LazyItemScope for itemContent param of Lazy lists
Modifier.fillMaxSize() can't work inside the scrolling layout like LazyColumn/Row as the items are measured with Constraints.Infinity as the constraints for the main axis. The new scope provides new Modifier.fillParentMaxSize() and similar methods for width and height to allow to support the use case when the item should fill the viewport.
Modifier.fillMaxSize() and similar are also added into the scope just to shadow the real methods and they are marked as deprecated. So once the users will try to use them the deprecated replace-to mechanism will guide them to use correct modifiers.
We didn't just use Modifier.fillMaxSize() as this could be too magical and the intended behaviour could be lost when you slightly refactor the code and are not within the scope anymore

Fixes: 162248854
Relnote: LazyItemScope was added for itemContent param of Lazy lists. It provides modifiers to fill the parent max size which solves the use case when the item should fill the viewport and the regular Modifier.fillMaxSize() doesn't work as the item is measured with infinity constraints.
Test: new tests, new demo for manual testing
Change-Id: Ibd3b21685641c22f7deaab1bb71785d8d6135058