Fix for previews not resizing if the component needs to grow
After optimization in aosp/1316677 we stopped calling requestLayout(), but there is a corner case when we still need it. Because the measuring system in Android Views is allowing to measure a child multiple times during one measure pass it is possible that the constraints we have are not the real constraints our parent has. For example some layout types first measure the child with the real constraints (say 1000x500), and the child responds that it only needs (70x70). Then the parent measures the child again with exactly the size child asked (70x70). Now our compose view thinks there is no possibility for us to fill the size larger that 70x70 and when the compose content will ask to be remeasured to 80x80 we will think it is not possible. But if instead of measuring using the constraints we have we call requestRelayout() on a view the parent will again measure us with 1000x500 first so can take 80x80.
In this cl I added a logic to detect this case (we were measured twice with different constraints during the same measure pass) and if it is true and we want to remeasure which can theoretically affect a root node we will call requestLayout(). It shouldn't affect the simpler cases when the parent is measuring us only once in terms of performance.

Fixes: 160391362
Test: covered this case with a new test
Change-Id: Ib4c6e18b424eea7d691464e4946b2591481ee74c