Wait for initial animations before continuing tests
When using FragmentStateManager, animations can
be cancelled or combined even when not reordering.
By waiting for the intial animation to complete, we
ensure that the replace animation later runs in
isolation.

We check the nullability of child views immediately
after kicking off the parent transaction since
Animation's waitforEnd() is based on sleep calls
which do not accurately catch the end before the
onDestroyView() happens synchronously with the
Animation ending.

Test: tests pass now
BUG: 147749580
Change-Id: Ib1572a82e70e62932482cae0ea9fc1e9ca704791