Support INITIALIZING in setMaxLifecycle()
setMaxLifecycle() currently does not support setting the max lifecycle to
INITIALIZING. This means it is not possible to hold a fragment in the
attached state, it will always go through onCreate().

As long as the Fragment has not yet been CREATED, we should allow its
maxLifecycle to be set to INITIALIZING.

Since fragments that have a max state of INITIALIZING may never reach
created, they cannot be allowed to access ViewModels until they are
created. This prevents the creation of ViewModels that may never be
destroyed since fragment clear view models in onDestroy().

Test: adjusted and added BackStackStateTest
Bug: 159662173
Change-Id: I63e48863ccce737d6427756d62e0e32d6c7d8675