Allow having multiple value providers for the same savedInstanceState() key
Originally we were hoping to solve the unique keys issue fully on a compose compiler/runtime side, but it is not trivial and will make the runtime way less efficient still without 100 percent uniqueness guarantees.

The simple case which is not currently working is
list.forEach {
   savedInstanceState { 0 }
}
Runtime keys are the keys generated during the compilation for the hash of the function invocation location in the codebase. In the for loop we have two invocations with the same key and we can't do anything more clever than just allow multiple invocations with the same key and use the order in which they are executed to differentiate them. Which makes this code basically equivalent to
list.forEachIndexed { index, item ->
   key(index) {
       savedInstanceState { 0 }
   }
}
But still allows to provide more reasonable keys if the user needs it
list.forEach { item ->
   key(item.myUniqueKey) {
       savedInstanceState { 0 }
   }
}
To allow state to be "moved" together with the item if the items in a list was reordered.

Given that we have to adjust how UiSavedStateRegistry works and replicate the runtime behavior. There are could be multiple providers registered for the same [key]. In this case the order in which they were registered matters.

Say we registered two providers for the key. One provides "1", second provides "2".
[performSave] in this case will have listOf("1", "2) as a value for the key in the map. And later, when the registry will be recreated with the previously saved values, the first execution of [consumeRestored] would consume "1" and the second one "2".

In addition to this change I also refactored rememberSavedInstanceState() implementation to be more prepared for future possible cases when the compose transaction can fail (we can only save the values during composition in state objects) and should use onCommit like behaviour for registering/unregistering the providers.
Also in light of deprecation of state { } composable in favour to remember { mutableStateOf() } I decided to add a more meaningful error message if the user will try to write rememberSavedInstanceState { mutableStateOf() } to guide people to use savedInstanceState() which provides a Saver implementation for MutableState

Fixes: 160042650
Fixes: 156853976
Fixes: 159026663
Fixes: 154920561
Relnote: Crash when something which saves the state was used inside the for loop is fixed. Now having the same key in savedInstanceState() is allowed, api of UiSavedStateRegistry is now adjusted to this new requirement
Test: new tests, manually
Change-Id: I4ab7630120ffce145d1bf09d52a475d197030150