Fix ExtensionTest and safely close ProcessingImageReader
Made sure that when the ProcessingImageReader is closed that the input
ImageReaderProxy and the ImageProxyBundle is not closed if the
CaptureProcessor is currently accessing it. This ensures that there
won't be any errors causes by the CaptureProcessor accessing an already
closed ImageProxy.

Also increased the timeout for
ExtensionTest.testCanBindToLifeCycleAndTakePicture since the
implementation of HDR extension can take extra long on some devices.

Bug: b/160793550, b/160612464
Test: ExtensionTest, ProcessingImageReaderTest, and all the
ImageCaptureTests

Change-Id: I0d3256e22849d1b0e837eb0e7326f551a6bdf448