Run scheduleRemoveTargets directly against shared element
When using AndroidX Transition 1.2.0, transition
listeners added to a TransitionSet of TransitionSet
do not get run. By changing the
scheduleRemoveTargets() to run directly against the
shared element transition in the same way as we
do for enter/exit transitions, we can verify that the
listener gets called on all versions of AndroidX
Transitions.

Test: manually testing in sample app with 1.2.0, 1.3.1
BUG: 162810352
Change-Id: I235d63427e6c3311bfda7ae530711d480f9bbea1