spacedBy and aligned arrangements for Row/Column
Relnote: The spacedBy Arrangement was added for Row and Column, to enable positioning layout children with a fixed spacing. The aligned Arrangment was also added, to enable positioning layout children one next to the other and aligned in the Row/Column according to an Alignment. The previous Arrangement.Vertical#arrange and Arrangement.Horizontal#arrange methods were deprecated, and writing custom Arrangements will not be supported in the future.
Fixes: 161985975
Fixes: 144976975
Test: RowColumnTest
Change-Id: I6733d4a76e4abd5713bfacf95da696bdd6153932