[GH] Fix OUT_DIR default to path that exists within GH checkout
Previously, `OUT_DIR` defaulted to a path that exists outside of the root folder for github checkout (it assumes this repo is nested in frameworks/support). This allows playground projects to ensure the OUT_DIR is writable when it is not specified.

Test: cd room && ./gradlew bOS

This is an imported pull request from https://github.com/androidx/androidx/pull/58.

Resolves #58
Github-Pr-Head-Sha: eadd96af502fe568e58ad334612bdd1e6950fa18
GitOrigin-RevId: 27a788e1ad4e5766ff363bbec72cc45da2ade86e
Change-Id: Ifbd054cdee485072a78a922259b40245323fd278