Expand kotlin metadata
This CL adds ability to parse properties and types in kotlin
metadata along with their nullability.

This is a preliminary work to start properly reading type
nullability out of kotlin code in Room.

I've also updated the test class to use actual test sources
rather than sources in the test itself as we already have the
kotlin compile testing setup in this project.

Bug: 163151422
Test: KotlinMetadataElementTest
Change-Id: I4ac3f4ddd4e14710119d84d5fac3094e75acb1ce