MediaRouter: correct onCreateRouteController(routeId, groupId)
When MediaRouter2ProviderServiceAdapter#onCreateSession is called,
it created member route controllers even before a group route is created.
As a result, wrong group id was sent to the provider.

This CL fixes that by defer creation of member route controllers.
It also fixes the sample app such that it won't work when the given group ID is incorrect.

Bug: 164984860
Test: Manually run support v7 demos w/ and w/o setSupportsDynamicGroupRoute
Change-Id: Ie5573c03717989a5feecb4c7d021f245d1dd5583